refactor(main): remove unnecessary "use client" directives
Convert performance-navbar to server component and delete duplicate
category-client hook in favor of passing server-fetched categories
as props. Rename category-server to category for consistency.
