+++ apps/main/messages/pt.po
@@ -2220,100 +2220,20 @@ msgctxt "ypy024"
 msgid "Tests your understanding with questions. Uses new scenarios to check if you grasped the concept, not just memorized it."
 msgstr "Testa sua compreensão com perguntas. Usa cenários novos pra verificar se você entendeu o conceito, não só memorizou ele."
 
-#: src/lib/categories/category-client.ts
+#: src/lib/categories/category-server.ts
+msgctxt "4HKnmB"
+msgid "{category} Courses"
+msgstr "Cursos de {category}"
+
 #: src/lib/categories/category-server.ts
 msgctxt "djJp6c"
 msgid "History"
 msgstr "História"
 
-#: src/lib/categories/category-client.ts
 #: src/lib/categories/category-server.ts
 msgctxt "dPgwrL"
 msgid "Math"
 msgstr "Matemática"
-
-#: src/lib/categories/category-client.ts
-#: src/lib/categories/category-server.ts
-msgctxt "fFt3il"
-msgid "Arts"
-msgstr "Artes"
-
-#: src/lib/categories/category-client.ts
-#: src/lib/categories/category-server.ts
-msgctxt "GsBRWL"
-msgid "Languages"
-msgstr "Idiomas"
-
-#: src/lib/categories/category-client.ts
-#: src/lib/categories/category-server.ts
-msgctxt "hlmkcL"
-msgid "Health"
-msgstr "Saúde"
-
-#: src/lib/categories/category-client.ts
-#: src/lib/categories/category-server.ts
-msgctxt "hwL7a0"
-msgid "Engineering"
-msgstr "Engenharia"
-
-#: src/lib/categories/category-client.ts
-#: src/lib/categories/category-server.ts
-msgctxt "I1ZCPK"
-msgid "Technology"
-msgstr "Tecnologia"
-
-#: src/lib/categories/category-client.ts
-#: src/lib/categories/category-server.ts
-msgctxt "KOuDOE"
-msgid "Economics"
-msgstr "Economia"
-
-#: src/lib/categories/category-client.ts
-#: src/lib/categories/category-server.ts
-msgctxt "mn6pIm"
-msgid "Culture"
-msgstr "Cultura"
-
-#: src/lib/categories/category-client.ts
-#: src/lib/categories/category-server.ts
-msgctxt "n5D/Ku"
-msgid "Communication"
-msgstr "Comunicação"
-
-#: src/lib/categories/category-client.ts
-#: src/lib/categories/category-server.ts
-msgctxt "OV/2+4"
-msgid "Society"
-msgstr "Sociedade"
-
-#: src/lib/categories/category-client.ts
-#: src/lib/categories/category-server.ts
-msgctxt "qydxOd"
-msgid "Science"
-msgstr "Ciências"
-
-#: src/lib/categories/category-client.ts
-#: src/lib/categories/category-server.ts
-msgctxt "u3ovjm"
-msgid "Law"
-msgstr "Direito"
-
-#: src/lib/categories/category-client.ts
-#: src/lib/categories/category-server.ts
-msgctxt "w1Fanr"
-msgid "Business"
-msgstr "Negócios"
-
-#: src/lib/categories/category-client.ts
-#: src/lib/categories/category-server.ts
-msgctxt "YMDBNH"
-msgid "Geography"
-msgstr "Geografia"
-
-#: src/lib/categories/category-server.ts
-msgctxt "4HKnmB"
-msgid "{category} Courses"
-msgstr "Cursos de {category}"
 
 #: src/lib/categories/category-server.ts
 msgctxt "EMi23I"
@@ -2321,11 +2241,76 @@ msgid "Explore all {category} courses"
 msgstr "Explorar todos os cursos de {category}"
 
 #: src/lib/categories/category-server.ts
+msgctxt "fFt3il"
+msgid "Arts"
+msgstr "Artes"
+
+#: src/lib/categories/category-server.ts
+msgctxt "GsBRWL"
+msgid "Languages"
+msgstr "Idiomas"
+
+#: src/lib/categories/category-server.ts
+msgctxt "hlmkcL"
+msgid "Health"
+msgstr "Saúde"
+
+#: src/lib/categories/category-server.ts
+msgctxt "hwL7a0"
+msgid "Engineering"
+msgstr "Engenharia"
+
+#: src/lib/categories/category-server.ts
+msgctxt "I1ZCPK"
+msgid "Technology"
+msgstr "Tecnologia"
+
+#: src/lib/categories/category-server.ts
+msgctxt "KOuDOE"
+msgid "Economics"
+msgstr "Economia"
+
+#: src/lib/categories/category-server.ts
+msgctxt "mn6pIm"
+msgid "Culture"
+msgstr "Cultura"
+
+#: src/lib/categories/category-server.ts
+msgctxt "n5D/Ku"
+msgid "Communication"
+msgstr "Comunicação"
+
+#: src/lib/categories/category-server.ts
+msgctxt "OV/2+4"
+msgid "Society"
+msgstr "Sociedade"
+
+#: src/lib/categories/category-server.ts
+msgctxt "qydxOd"
+msgid "Science"
+msgstr "Ciências"
+
+#: src/lib/categories/category-server.ts
 msgctxt "Rz9oL+"
 msgid "{category} courses"
 msgstr "Cursos de {category}"
 
 #: src/lib/categories/category-server.ts
+msgctxt "u3ovjm"
+msgid "Law"
+msgstr "Direito"
+
+#: src/lib/categories/category-server.ts
+msgctxt "w1Fanr"
+msgid "Business"
+msgstr "Negócios"
+
+#: src/lib/categories/category-server.ts
 msgctxt "XL8IWT"
 msgid "{category} courses. The best {category} online courses, interactive lessons to learn."
 msgstr "Cursos de {category}. Os melhores cursos online de {category}, aulas interativas pra aprender."
+
+#: src/lib/categories/category-server.ts
+msgctxt "YMDBNH"
+msgid "Geography"
+msgstr "Geografia"
