+++ apps/main/src/app/(catalog)/courses/layout.tsx
@@ -1,11 +1,17 @@
+import { getCategories } from "@/lib/categories/category";
 import { Suspense } from "react";
 import { CategoryPills, CategoryPillsSkeleton } from "./category-pills";
 
-export default async function CoursesLayout({ children }: { children: React.ReactNode }) {
+async function CategoryPillsWithData() {
+  const categories = await getCategories();
+  return <CategoryPills categories={categories} />;
+}
+
+export default async function CoursesLayout({ children }: LayoutProps<"/courses">) {
   return (
     <>
       <Suspense fallback={<CategoryPillsSkeleton />}>
-        <CategoryPills />
+        <CategoryPillsWithData />
       </Suspense>
 
       {children}
